feat: support binance usd margined endpoints

+++ binance/spot/marketdata/types/orderbook.go
@@ -6,7 +6,7 @@ type GetOrderbookParams struct {
 }
 
 type Orderbook struct {
-	LastUpdateID int        `json:"lastUpdateId"`
+	LastUpdateID int64        `json:"lastUpdateId"`
 	Bids         [][]string `json:"bids"`
 	Asks         [][]string `json:"asks"`
 }
